Split long e2e tests
